requirements.txt
pip
numpy
scipy
matplotlib
pandas
scikit-learn
seaborn
tensorflow
typing-extensions==4.6.2
sentencepiece==0.1.99
pytorch_lightning==1.9.5
tqdm==4.65.0
transformers==4.16.0
openai==0.27.7
Requests==2.31.0
sentence-transformers
easydict
editdistance